+++ Project/BarrocIntens/App.xaml
@@ -8,9 +8,9 @@
         <ResourceDictionary>
             <ResourceDictionary.MergedDictionaries>
                 <XamlControlsResources xmlns="using:Microsoft.UI.Xaml.Controls" />
-                <ResourceDictionary Source="ResourceDictionary1.xaml" />
+                <!-- Other merged dictionaries here -->
             </ResourceDictionary.MergedDictionaries>
-
+            
         </ResourceDictionary>
     </Application.Resources>
 </Application>
