+++ Project/BarrocIntens/BarrocIntens.csproj
@@ -21,10 +21,8 @@
   </ItemGroup>
   <ItemGroup>
     <None Remove="Assets\meldingIcon.png" />
-    <None Remove="Financiën\FinanciënMainWindow.xaml" />
     <None Remove="Onderhoud\OnderhoudBaseWindow.xaml" />
     <None Remove="Onderhoud\OnderhoudIngekomenStoringen.xaml" />
-    <None Remove="ResourceDictionary1.xaml" />
     <None Remove="SalesDashboard.xaml" />
   </ItemGroup>
 
@@ -133,14 +131,12 @@
 
   <ItemGroup>
     <Page Update="Sales\SalesEditNotePage.xaml">
-    <Page Update="ResourceDictionary1.xaml">
       <Generator>MSBuild:Compile</Generator>
     </Page>
   </ItemGroup>
 
   <ItemGroup>
     <Page Update="Sales\SalesCreateNotePage.xaml">
-    <Page Update="Financiën\FinanciënMainWindow.xaml">
       <Generator>MSBuild:Compile</Generator>
     </Page>
   </ItemGroup>
