+++ Project/BarrocIntens/Data/Department.cs
@@ -1,4 +1,4 @@
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
@@ -6,10 +6,9 @@ using System.Threading.Tasks;
 
 namespace BarrocIntens.Data
 {
-    public class ProductCategories
+    internal class Department
     {
         public int Id { get; set; }
         public string Name { get; set; }
-        public ICollection<Products> Products { get; set; }
     }
 }
