+++ Project/BarrocIntens/Data/Product.cs
@@ -1,4 +1,4 @@
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
@@ -6,7 +6,7 @@ using System.Threading.Tasks;
 
 namespace BarrocIntens.Data
 {
-    public class Products
+    internal class Product
     {
         public int Id { get; set; }
         public string Name { get; set; }
