--- Project/BarrocIntens/Financiën/FinanciënMainWindow.xaml
@@ -1,59 +1,0 @@
-﻿<?xml version="1.0" ?>
-<Window
-    x:Class="BarrocIntens.Financiën.FinanciënMainWindow"
-    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:local="using:BarrocIntens.Financiën"
-    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-    mc:Ignorable="d"
-    >
-
-    
-    <Grid Background="#3e3e3e">
-        
-        <!-- Titel bovenaan -->
-        <TextBlock Text="Financiën" Foreground="#FDD716" FontSize="40" FontWeight="Bold" HorizontalAlignment="Center" VerticalAlignment="Top" Margin="0,20,0,0"/>
-
-        <!-- Linker zijpaneel met knoppen -->
-        <StackPanel VerticalAlignment="Center" HorizontalAlignment="Left" Margin="40">
-            <Image Source="assets/Logo6_groot.png" Width="100" Height="100" Margin="0,0,0,20"/>
-            <Button Content="Klant Gegevens" Style="{StaticResource HoverButtonStyle}"/>
-            <Button Content="Contract Aanmaken" Style="{StaticResource HoverButtonStyle}"/>
-            <Button Content="Factuur Aanmaken" Style="{StaticResource HoverButtonStyle}"/>
-            <Button Content="Maand/Jaar Opgave" Style="{StaticResource HoverButtonStyle}"/>
-        </StackPanel>
-
-        <!-- Hoofdcontent in het midden -->
-        <StackPanel HorizontalAlignment="Center" VerticalAlignment="Center" Orientation="Vertical" Margin="30">
-            <Border Width="600" Background="Black" CornerRadius="20,20,20,20" Padding="0" BorderBrush="#FDD716" BorderThickness="3">
-                <StackPanel>
-                    <!-- Full-width Title Bar -->
-                    <Border Background="#FDD716" CornerRadius="10,10,0,0" Padding="10">
-                        <TextBlock Text="Contracten" Foreground="Black" FontSize="26" FontWeight="Bold" HorizontalAlignment="Center"/>
-                    </Border>
-
-                    <!-- Kolom headers -->
-                    <StackPanel Orientation="Horizontal" Margin="0,10" Padding="10,0,10,0">
-                        <TextBlock Text="Naam" Foreground="#FDD716" FontSize="20" FontWeight="Bold" Width="200" TextAlignment="Center"/>
-                        <TextBlock Text="Start Datum" Foreground="#FDD716" FontSize="20" FontWeight="Bold" Width="200" TextAlignment="Center"/>
-                        <TextBlock Text="Eind Datum" Foreground="#FDD716" FontSize="20" FontWeight="Bold" Width="200" TextAlignment="Center"/>
-                    </StackPanel>
-
-                    <Rectangle Fill="Black" Height="2" Margin="0,5"/>
-                    <ItemsControl ItemsSource="{Binding ContractLijst}">
-                        <ItemsControl.ItemTemplate>
-                            <DataTemplate>
-                                <StackPanel Orientation="Horizontal" Margin="0,10" Padding="10,0,10,0">
-                                    <TextBlock Text="{Binding KlantNaam}" Foreground="#FDD716" FontSize="18" Width="200" TextAlignment="Center"/>
-                                    <TextBlock Text="{Binding StartDatumFormatted}" Foreground="#FDD716" FontSize="18" Width="200" TextAlignment="Center"/>
-                                    <TextBlock Text="{Binding EindDatumFormatted}" Foreground="#FDD716" FontSize="18" Width="200" TextAlignment="Center"/>
-                                </StackPanel>
-                            </DataTemplate>
-                        </ItemsControl.ItemTemplate>
-                    </ItemsControl>
-                </StackPanel>
-            </Border>
-        </StackPanel>
-    </Grid>
-</Window>
