+++ Project/BarrocIntens/Inkoop/ProductAanmaakPage.xaml.cs
@@ -1,4 +1,3 @@
-using BarrocIntens.Data;
 using Microsoft.UI.Xaml;
 using Microsoft.UI.Xaml.Controls;
 using Microsoft.UI.Xaml.Controls.Primitives;
@@ -13,7 +12,6 @@ using System.Linq;
 using System.Runtime.InteropServices.WindowsRuntime;
 using Windows.Foundation;
 using Windows.Foundation.Collections;
-using static System.Runtime.InteropServices.JavaScript.JSType;
 
 // To learn more about WinUI, the WinUI project structure,
 // and more about our project templates, see: http://aka.ms/winui-project-info.
@@ -28,13 +26,6 @@ namespace BarrocIntens.Inkoop
         public ProductAanmaakPage()
         {
             this.InitializeComponent();
-
-            using (var db = new AppDbContext())
-            {
-                var categories = db.ProductCategories.ToList();
-                CategoryComboBox.ItemsSource = categories;
-                CategoryComboBox.SelectedIndex = 0;
-            }
         }
 
         private void TerugButton_Click(object sender, RoutedEventArgs e)
@@ -44,67 +35,7 @@ namespace BarrocIntens.Inkoop
 
         private void OpslaanButton_Click(object sender, RoutedEventArgs e)
         {
-            NaamError.Visibility = Visibility.Collapsed;
-            DescError.Visibility = Visibility.Collapsed;
-            PrijsError.Visibility = Visibility.Collapsed;
-            CategoryError.Visibility = Visibility.Collapsed;
 
-            int validatieErrors = 0;
-
-            if (NaamInput.Text.Length == 0)
-            { 
-                NaamError.Visibility = Visibility.Visible;
-                validatieErrors += 1;
-            }
-
-            if (DescInput.Text.Length == 0)
-            {
-                DescError.Visibility = Visibility.Visible;
-                validatieErrors += 1;
-            }
-
-
-            if (decimal.TryParse(PrijsInput.Text, out decimal prijsOutput))
-            {
-
-            }
-            else
-            {
-                PrijsError.Visibility= Visibility.Visible;
-                validatieErrors += 1;
-            }
-
-            if (CategoryComboBox.SelectedValue == null)
-            {
-                CategoryError.Visibility = Visibility.Visible;
-                validatieErrors += 1;
-            }
-
-            prijsOutput = Math.Round(prijsOutput, 2);
-
-            if (validatieErrors == 0) 
-            {
-                using (var db = new AppDbContext())
-                {
-                    db.Products.Add(new Products
-                    {
-                        Name = NaamInput.Text,
-                        Description = DescInput.Text,
-                        Price = prijsOutput,
-                        IsStock = VoorraadCheckBox.IsChecked == true,
-                        VisibleForCustomers = ZichtbaarheidCheckBox.IsChecked == true,
-                        CategoryId = (int)CategoryComboBox.SelectedValue
-                    });
-
-                    db.SaveChanges();
-                    Frame.Navigate(typeof(ProductenPage));
-                    return;
-                }
-            }
-            else
-            {
-                return;
-            }
         }
     }
 }
