+++ Project/BarrocIntens/Inkoop/ProductenPage.xaml
@@ -5,8 +5,7 @@
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:local="using:BarrocIntens.Inkoop"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
-    xmlns:localData="using:BarrocIntens.Data"
+    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
     mc:Ignorable="d"
     Background="{ThemeResource ApplicationPageBackgroundThemeBrush}">
 
@@ -23,25 +22,8 @@
                             <TextBox x:Name="SearchTextBox" PlaceholderText="Zoek hier.." Width="650"/>
                             <Button Click="ZoekButton_Click" Background="Black" Foreground="#ffd700" BorderBrush="#ffd700" BorderThickness="2">Zoek</Button>
                         </StackPanel>
-                        <StackPanel MaxHeight="400" MaxWidth="700">
-                            <!-- Producten ListView -->
-                            <ListView x:Name="ProductListView" Height="400" Width="700" Margin="0 10 0 0">
-                                <ListView.ItemTemplate>
-                                    <DataTemplate x:DataType="localData:Products">
-                                        <StackPanel BorderBrush="#ffd700" BorderThickness="2" Margin="0 5 10 5">
-                                            <StackPanel Margin="10 5 5 5">
-                                                <TextBlock FontSize="25" Foreground="#ffd700" Text="{x:Bind Name}" TextWrapping="Wrap" Margin="0 0 0 10"/>
-                                                <TextBlock FontSize="15" Foreground="#ffd700" Text="{x:Bind Description}" TextWrapping="Wrap"/>
-                                                <TextBlock FontSize="15" Foreground="#ffd700" Text="{x:Bind Category.Name, FallbackValue='No Category'}" TextWrapping="Wrap" Margin="0 10 0 10"/>
-                                            </StackPanel>
-                                            <StackPanel Orientation="Horizontal" Spacing="40" Margin="270 10 20 15">
-                                                <Button Click="BewerkButton_Click" Background="Black" Foreground="#ffd700" BorderBrush="#ffd700" BorderThickness="2" Width="150" FontSize="15" HorizontalAlignment="Center">Bewerken</Button>
-                                                <Button Click="VerwijderButton_Click" Background="Black" Foreground="#ffd700" BorderBrush="#ffd700" BorderThickness="2" Width="150" FontSize="15" HorizontalAlignment="Center">Verwijderen</Button>
-                                            </StackPanel>
-                                        </StackPanel>
-                                    </DataTemplate>
-                                </ListView.ItemTemplate>
-                            </ListView>
+                        <StackPanel Height="400">
+                            <!--Plaats hier de Producten Listview-->
                         </StackPanel>
                     </StackPanel>
                 </StackPanel>
