+++ Project/BarrocIntens/Inkoop/ProductenPage.xaml.cs
@@ -1,5 +1,3 @@
-using BarrocIntens.Data;
-using Microsoft.EntityFrameworkCore;
 using Microsoft.UI.Xaml;
 using Microsoft.UI.Xaml.Controls;
 using Microsoft.UI.Xaml.Controls.Primitives;
@@ -9,11 +7,9 @@ using Microsoft.UI.Xaml.Media;
 using Microsoft.UI.Xaml.Navigation;
 using System;
 using System.Collections.Generic;
-using System.Collections.ObjectModel;
 using System.IO;
 using System.Linq;
 using System.Runtime.InteropServices.WindowsRuntime;
-using System.Threading.Tasks;
 using Windows.Foundation;
 using Windows.Foundation.Collections;
 
@@ -30,12 +26,6 @@ namespace BarrocIntens.Inkoop
         public ProductenPage()
         {
             this.InitializeComponent();
-
-            using (var db = new AppDbContext())
-            {
-                ProductListView.ItemsSource = db.Products.Include(p => p.Category).OrderBy(p => p.Id).ToList();
-            };
-           
         }
 
         private void ZoekButton_Click(object sender, RoutedEventArgs e)
@@ -52,53 +42,5 @@ namespace BarrocIntens.Inkoop
         {
             Frame.Navigate(typeof(ProductAanmaakPage));
         }
-
-        private void BewerkButton_Click(object sender, RoutedEventArgs e)
-        {
-
-        }
-
-        private async void VerwijderButton_Click(object sender, RoutedEventArgs e)
-        {
-            var button = sender as Button;
-            if (button != null)
-            {
-                var product = button.DataContext as Products;
-
-                if (product != null)
-                {
-                    bool isDeleted = await DeleteProductFromDatabase(product);
-
-                    if (isDeleted)
-                    {
-                        var productList = ProductListView.ItemsSource as ObservableCollection<Products>;
-                        if (productList != null)
-                        {
-                            productList.Remove(product);
-                        }
-                    }
-                }
-            }
-        }
-
-        private async Task<bool> DeleteProductFromDatabase(Products product)
-        {
-            using (var db = new AppDbContext())
-            {
-                var productToDelete = await db.Products
-                                                    .FirstOrDefaultAsync(p => p.Id == product.Id);
-
-                if (productToDelete != null)
-                {
-                    db.Products.Remove(productToDelete);
-                    await db.SaveChangesAsync();
-                    ProductListView.ItemsSource = db.Products.Include(p => p.Category).OrderBy(p => p.Id).ToList();
-                    return true;
-                }
-            }
-
-            return false;
-        }
-
     }
 }
