+++ Project/BarrocIntens/LoginWindow.xaml.cs
@@ -65,9 +65,9 @@ namespace BarrocIntens
 					else if (departmentId == 3)
 					{
 						// Verander dit naar de juiste window wanneer deze is aangemaakt
-						var financeDashboard = new Financiën.FinanciënMainWindow();
+						var financeDashboard = new Financiën.FinanciënMainPage();
 						this.Close();
-						financeDashboard.Activate();
+						//financeDashboard.Activate();
 					}
 					else if (departmentId == 4)
 					{
